+++ test-support/helpers/ember-formulaic/test-helpers.js
@@ -1,0 +1,4 @@
+import '../ember-i18n/test-helpers';
+
+import 'ember-formulaic/fill-form';
+import 'ember-formulaic/click-on';