+++ test-support/helpers/test-helpers.js
@@ -1,4 +1,0 @@
-import './ember-i18n/test-helpers';
-
-import './fill-form';
-import './click-on';