+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 master
 ------
 
+* Restructure files to enable `import './ember-formulaic/test-helpers';`
 * Change behvaior of `fillForm` to only work on
   `<label for="text">Text</label><input id="text">`, in addition to
   `<label>Text <input></label>`